+++ .vscode/settings.json
@@ -1,0 +1,6 @@
+{
+  "editor.tabSize": 4,
+  "editor.detectIndentation": false,
+"editor.insertSpaces": true,
+"rust-analyzer.check.command": "clippy"
+}
